Revit 2024 (#219)
* Add support for Revit 2024

* Create "auto-upgrade" script to automate portions of adding support for a new Revit version

* ElementId int -> long changes in Revit 2024 API

* fix(Room to Mass): Area names not including trailing numbers

* feat(RoomsToMass): Allow masses at centerline of walls, area name fix

---------

+++ AVF Manager/src/HOK.ViewAnalysis/HOK.ViewAnalysis/MainWindow.xaml.cs
@@ -95,7 +95,7 @@ namespace HOK.ViewAnalysis
             }
         }
 
-        private bool SaveResultData(Dictionary<int,RoomData> roomDictionary)
+        private bool SaveResultData(Dictionary<long,RoomData> roomDictionary)
         {
             bool saved = false;
             try
